+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  * Support for synthetic initializers:
    * Full Parameter information in the IDE is supported if the `__init__` method is re-declared.
    * Otherwise, support is limited to argument name completion.
+ * Full Parameter information for getters and setters in the IDE.
 
 ## 0.2.0 - 2016-11-05